feat: [#10] 컴포넌트 재사용을 위한 데이터 fetch 파일 추가
